Maximize Grid Happiness (incomplete)
This solves the problem using a genetic algorithm. It works after a fashion. On
the LeetCode platform, it times out. However, it passes all the test cases I am
aware of. It took some tuning of the hyperparameters to get it to this point.
In the process, when it produced an incorrect response, it was due to finding a
local maxima, which was almost always only 10 less than the global maximum.

https://leetcode.com/problems/maximize-grid-happiness/